Hash authentication through ilias plugin
This commit allows to the annotation tool to verify if a logged
user comes from ilias and from the correct course. If not, It
will block the access to the user. Also, verifies if the user is
a course administrator or a normal user. This is useful for the
annotation rights management.